requirements.txt
pandas
numpy
streamlit
torch
torchvision
pillow
opencv-python
pickle